requirements.txt
dotenv==0.9.9
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-openai==0.3.6
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.9
Markdown==3.7
openai==1.63.2
pandas==2.2.3
Flask==3.1.0
pymongo==4.11.2
reportlab==4.3.1
identity==0.9.2
loguru==0.7.3
opencensus-ext-azure==1.1.14
pydantic==2.10.6